+++ src/main/java/com/soocompany/wodify/box/domain/Box.java
@@ -2,7 +2,7 @@ package com.soocompany.wodify.box.domain;
 
 
 
-import com.soocompany.wodify.common.BaseEntity;
+import com.soocompany.wodify.common.domain.BaseEntity;
 import com.soocompany.wodify.member.domain.Member;
 import lombok.*;
 import javax.persistence.*;
