+++ src/main/java/com/soocompany/wodify/category/domain/Category.java
@@ -1,6 +1,6 @@
 package com.soocompany.wodify.wod.category.domain;
 
-import com.soocompany.wodify.common.BaseEntity;
+import com.soocompany.wodify.common.domain.BaseEntity;
 import com.soocompany.wodify.wod.category.dto.CategoryListResDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
