+++ src/main/java/com/soocompany/wodify/common/dto/CommonResDto.java
@@ -1,4 +1,4 @@
-package com.soocompany.wodify.common.exception;
+package com.soocompany.wodify.common.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
