+++ src/main/java/com/soocompany/wodify/reservation/domain/Reservation.java
@@ -1,11 +1,10 @@
 package com.soocompany.wodify.reservation.domain;
 
 import com.soocompany.wodify.box.domain.Box;
-import com.soocompany.wodify.common.BaseEntity;
+import com.soocompany.wodify.common.domain.BaseEntity;
 import com.soocompany.wodify.member.domain.Member;
 import com.soocompany.wodify.reservation.dto.ReservationDetailResDto;
 import com.soocompany.wodify.reservation.dto.ReservationListResDto;
-import com.soocompany.wodify.reservation.repository.ReservationRepository;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
