+++ src/main/java/com/soocompany/wodify/reservation_detail/domain/ReservationDetail.java
@@ -1,7 +1,7 @@
 package com.soocompany.wodify.reservation_detail.domain;
 
 import com.soocompany.wodify.box.domain.Box;
-import com.soocompany.wodify.common.BaseEntity;
+import com.soocompany.wodify.common.domain.BaseEntity;
 import com.soocompany.wodify.member.domain.Member;
 import com.soocompany.wodify.reservation.domain.Reservation;
 import com.soocompany.wodify.reservation_detail.dto.ReservationDetailDetResDto;
